+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [1.2.2]
+- Cleaned up the `LocationFileManager` class
 
 ## [1.2.1] - 13/07/2020
 ### Changed